% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm-package.R
\docType{package}
\name{ADMM}
\alias{ADMM}
\alias{ADMM-package}
\title{ADMM : Algorithms using Alternating Direction Method of Multipliers}
\description{
An introduction of Alternating Direction Method of Multipliers (ADMM) method has been a breakthrough in
solving complex and non-convex optimization problems in a reasonably stable as well as scalable fashion.
Our package aims at providing handy tools for fast computation on well-known problems using the method.
For interested users/readers, please visit Prof. Stephen Boyd's \href{http://stanford.edu/~boyd/papers/admm_distr_stats.html}{website}
entirely devoted to the topic.
}
