% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{aux_genetinversion}
\alias{aux_genetinversion}
\title{inversion : use half of the cores}
\usage{
aux_genetinversion(A, rho, L, R, lambda2, parallel = FALSE,
  nCore = ceiling(detectCores()/2))
}
\description{
inversion : use half of the cores
}
