% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{mat_shrink}
\alias{mat_shrink}
\title{Soft threshholding operator.}
\usage{
mat_shrink(K, tau)
}
\arguments{
\item{K}{matrix}

\item{tau}{regularization parameter}
}
\value{
Matrix
}
\description{
Applies the shrinkage operator for singular value tresholding.
}
