% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.R
\name{matrix_plantsubm}
\alias{matrix_plantsubm}
\title{Sample matrix}
\usage{
matrix_plantsubm(M, N, c, n, p, q)
}
\arguments{
\item{M}{number of rows in sampled matrix}

\item{N}{number of rows in sampled matrix}

\item{c}{natural number used to calculate number of rows in dense submatrix}

\item{n}{natural number used to calculate number of rows in dense submatrix}

\item{p}{density outside planted submatrix}

\item{q}{density inside planted submatrix}
}
\value{
Matrix G sampled from the planted dense (mn)-submatrix model, dense sumbatrix X0, matrix Y0 used to count the number of disagreements between G and X0
}
\description{
Generates binary (M,N) - matrix sampled from dense (m,n) - submatrix
Let U* and V* be m and n index sets.
For each i in U*, j in V* we let a_ij = 1 with probability q and 0  otherwise.
For each remaining (ij) we set a_ij = 1 with probability p < q and take a_ij = 0 otherwise.
}
